+++ havokmud/act.wizard.c
@@ -6,6 +6,7 @@
 #include "config.h"
 #include <stdio.h>
 #include <string.h>
+#include <strings.h>
 #include <ctype.h>
 #include <time.h>
 #include <unistd.h>
@@ -1202,7 +1203,7 @@ void do_at(struct char_data *ch, char *argument, int cmd)
         return;
     }
     
-    if (isdigit(*loc_str)) {
+    if (isdigit((int)*loc_str)) {
         loc_nr = atoi(loc_str);
         if (NULL == real_roomp(loc_nr)) {
             send_to_char("No room exists with that number.\n\r", ch);
@@ -1265,7 +1266,7 @@ void do_goto(struct char_data *ch, char *argument, int cmd)
         send_to_char("You must supply a room number or a name.\n\r", ch);
         return;
     }
-    if (isdigit(*buf) && NULL == index(buf, '.')) {
+    if (isdigit((int)*buf) && NULL == index(buf, '.')) {
         loc_nr = atoi(buf);
         if (!real_roomp(loc_nr)) {
             if (GetMaxLevel(ch) < CREATOR || loc_nr < 0) {
@@ -3264,7 +3265,7 @@ void do_load(struct char_data *ch, char *argument, int cmd)
         return;
     }
 
-    if (isdigit(*num)) {
+    if (isdigit((int)*num)) {
         number = atoi(num);
     } else {
         number = -1;
@@ -3507,7 +3508,7 @@ void do_purge(struct char_data *ch, char *argument, int cmd)
             }
 
             argument = get_argument(argument, &name);
-            if (!name || !isdigit(*name)) {
+            if (!name || !isdigit((int)*name)) {
                 send_to_char("purge room start [end]", ch);
                 return;
             }
@@ -3515,7 +3516,7 @@ void do_purge(struct char_data *ch, char *argument, int cmd)
             range[0] = atoi(name);
 
             argument = get_argument(argument, &name);
-            if (name && isdigit(*name)) {
+            if (name && isdigit((int)*name)) {
                 range[1] = atoi(name);
             } else {
                 range[1] = range[0];
@@ -4110,7 +4111,7 @@ void do_advance(struct char_data *ch, char *argument, int cmd)
         send_to_char("You must supply a level number.\n\r", ch);
         return;
     } else {
-        if (!isdigit(*level)) {
+        if (!isdigit((int)*level)) {
             send_to_char("Third argument must be a positive integer.\n\r", ch);
             return;
         }
@@ -4562,7 +4563,7 @@ void do_show(struct char_data *ch, char *argument, int cmd)
             return;
         }
 
-        if( isdigit(*zonenum) ) {
+        if( isdigit((int)*zonenum) ) {
             zone = atoi( zonenum );
             if (zone <= 0 || zone > top_of_zone_table) {
                 send_to_char("That is not a valid zone_number\n\r", ch);
@@ -4605,7 +4606,7 @@ void do_show(struct char_data *ch, char *argument, int cmd)
                (which_i = obj_index, topi = top_of_objt)) {
 
         zonenum = skip_spaces(argument);
-        if (!zonenum || !(isdigit(*zonenum))) {
+        if (!zonenum || !(isdigit((int)*zonenum))) {
             append_to_string_block(&sb, "Usage:\n\r"
                                    "  show wearslot #\n\r"
                                    "  Number ranging from  0   TAKE\n\r"
@@ -4656,7 +4657,7 @@ void do_show(struct char_data *ch, char *argument, int cmd)
     } else if (is_abbrev(arg1, "itemtype") &&
                (which_i = obj_index, topi = top_of_objt)) {
         zonenum = skip_spaces(argument);
-        if (!zonenum || !(isdigit(*zonenum))) {
+        if (!zonenum || !(isdigit((int)*zonenum))) {
             append_to_string_block(&sb, "Usage:\n\r"
                                    "  show itemtype #\n\r"
                                    "  Number ranging from  0   UNDEFINED\n\r"
@@ -4720,7 +4721,7 @@ void do_show(struct char_data *ch, char *argument, int cmd)
         zonenum = skip_spaces(argument);
         zone = -1;
 
-        if( zonenum && isdigit(*zonenum) ) {
+        if( zonenum && isdigit((int)*zonenum) ) {
             zone = atoi(zonenum);
         }
 
@@ -4757,7 +4758,7 @@ void do_show(struct char_data *ch, char *argument, int cmd)
             return;
         }
 
-        if( isdigit(*zonenum) ) {
+        if( isdigit((int)*zonenum) ) {
             zone = atoi(zonenum);
         }
 
@@ -4808,7 +4809,7 @@ void do_show(struct char_data *ch, char *argument, int cmd)
 
         zonenum = skip_spaces(argument);
         zone = -1;
-        if( zonenum && isdigit(*zonenum) ) {
+        if( zonenum && isdigit((int)*zonenum) ) {
             zone = atoi(zonenum);
         }
 
@@ -4971,7 +4972,7 @@ void do_show(struct char_data *ch, char *argument, int cmd)
                (which_i = obj_index, topi = top_of_objt)) {
         zonenum = skip_spaces(argument);
         zone = -1;
-        if( zonenum && isdigit(*zonenum) ) {
+        if( zonenum && isdigit((int)*zonenum) ) {
             zone = atoi(zonenum);
         }
 
@@ -7275,7 +7276,7 @@ void do_setobjmax(struct char_data *ch, char *argument, int cmd)
     argument = get_argument(argument, &objec);
     num = skip_spaces(argument);
 
-    if (num && isdigit(*num)) {
+    if (num && isdigit((int)*num)) {
         number = atoi(num);
     } else {
         send_to_char("usage is: setobjmax itemname loadrate.\n\r", ch);
@@ -7328,7 +7329,7 @@ void do_setobjspeed(struct char_data *ch, char *argument, int cmd)
     argument = get_argument(argument, &objec);
     num = skip_spaces(argument);
 
-    if (num && isdigit(*num)) {
+    if (num && isdigit((int)*num)) {
         number = atoi(num);
     } else {
         send_to_char("usage is: setobjspeed item speed.\n\r", ch);
@@ -7370,7 +7371,7 @@ void do_setwtype(struct char_data *ch, char *argument, int cmd)
     argument = get_argument(argument, &objec);
     num = skip_spaces(argument);
 
-    if (num && isdigit(*num)) {
+    if (num && isdigit((int)*num)) {
         number = atoi(num);
     } else {
         send_to_char("Usage: setwtype item wtype.  Get wtype from allweapons "
