+++ havokmud/handler.c
@@ -38,7 +38,7 @@ char           *fname(char *namelist)
     static char     holder[30];
     register char  *point;
 
-    for (point = holder; isalpha(*namelist); namelist++, point++) {
+    for (point = holder; isalpha((int)*namelist); namelist++, point++) {
         *point = *namelist;
     }
     *point = '\0';
@@ -1581,7 +1581,7 @@ int get_number(char **name)
         strcpy(*name, ppos);
 
         for (i = 0; *(number + i); i++) {
-            if (!isdigit(*(number + i))) {
+            if (!isdigit((int)*(number + i))) {
                 return (0);
             }
         }
